Move sso token fn from local_xlsxBuilder into xlsxBuilder
